+++ src/ChipmunkSharp/Common/Helpers/RandomHelper.cs
@@ -11,9 +11,9 @@ namespace ChipmunkSharp
 
 		public static int RAND_MAX = 0x7fff;
 
-		public static float next(float min, float max)
+		public static double next(double min, double max)
 		{
-			return (float)((max - min) * random.NextDouble() + min);
+			return (double)((max - min) * random.NextDouble() + min);
 		}
 
 		public static int rand()
@@ -21,33 +21,33 @@ namespace ChipmunkSharp
 			return random.Next();
 		}
 
-		public static float randBell(float scale)
+		public static double randBell(double scale)
 		{
-			return (float)((scale) * (-(frand(.5f) + frand(.5f) + frand(.5f))));
+			return (double)((scale) * (-(frand(.5f) + frand(.5f) + frand(.5f))));
 		}
 
 
-		public static float frand() //HACK
+		public static double frand() //HACK
 		{
 			//double tmp = ((rand.NextDouble() *  f) / ((double) (/*(uint)~0*/ 0xFFFFFFFF /*or is it -1 :P */)));
 			//return tmp < 0 ? (-tmp) : tmp;
-			return (float)(random.NextDouble());
+			return (double)(random.NextDouble());
 		}
 
 		/// <summary>
 		/// This is bit spooky conversion of C -> C#...
 		/// </summary>
-		public static float frand(float f) //HACK
+		public static double frand(double f) //HACK
 		{
 			//double tmp = ((rand.NextDouble() *  f) / ((double) (/*(uint)~0*/ 0xFFFFFFFF /*or is it -1 :P */)));
 			//return tmp < 0 ? (-tmp) : tmp;
 			return frand() * f;
 		}
 
-		public static float FastDistance2D(float x, float y)
+		public static double FastDistance2D(double x, double y)
 		{
 			// this function computes the distance from 0,0 to x,y with ~3.5% error
-			float mn;
+			double mn;
 			// first compute the absolute value of x,y
 			x = (x < 0.0f) ? -x : x;
 			y = (y < 0.0f) ? -y : y;
